information_service/app/requirements.txt
Flask==2.1.2
Flask-RESTful==0.3.9
flask-mysql==1.5.2
APScheduler
certifi==2020.12.5
chardet==4.0.0
idna==2.10
msgpack==1.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
mysql-connector-python==8.0.29
#mysqlclient==2.1.0
#PyMySQL==1.0.2

